% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{state_rds}
\alias{state_rds}
\title{Capture an arbitrary R object}
\usage{
state_rds(obj)
}
\arguments{
\item{obj}{Any R object.}
}
\value{
An object of class `shinyswarm_rds`.
}
\description{
Marks an R object (e.g. a trained model) so that [capture_state()]
serialises it with [base::serialize()] and embeds it (base64-encoded) in the
checkpoint JSON. On restore the object is reconstructed with
[base::unserialize()].
}
\examples{
state_rds(lm(mpg ~ wt, data = mtcars))
}
